requirements.txt
pytesseract
opencv-python
pyzbar
numpy
pyqrcode
pypng
setuptools
mysql-connector-python-rf
pillow